Fix DAT-2307: Eliminate infinite recursion when deferred fields are present
When querying a django model you can use `.only(<field_name>)` or `.defer(<field_name>)`
to constrain the fields pulled from the db to just the fields requested instead of all of the
fields on the model. When/if those fields are accessed in the future then django
will make another trip to the database to request the "deferred" field. This is an optimization that
allows one to only get the fields they need for specific use cases.

In our previous implementation we were using `getattr`. In the case of deferred
fields if the field was deferred getattr would make a request out to the database
to get the field. Even in the case of field.attname. This would result in
infinite recursion where we would try to initialize an object using deferred fields,
then call getattr on those deferred fields, and then recurse trying to initialize
again. This bug didn't present itself because we rarely use .only/.defer in our
monolith code. In all cases where we do the models don't inheirt from ChangesMixin
so we never saw this bug.

Changing to use only the attributes already available on `__dict__` makes sure
we don't go out to the database and end up in infinite recursion. In the case of
deferred fields if the field wasn't explicitly requested then the field is not on
`__dict__` so we just won't include it in the fields returned by current_state().

In DAT-2307 this issue cropped up because django 3 optimized model delete
code to use a `.only()` (deferred field) under certain circumstances. This meant
that at times deletes would result in infinite recursion.